scout: overlay: Use oem_pocket sensor on keyguard
Our ultrasound proximity sensor does not work when the device is not
upright so falsing did not really work before. Additionally, enabling
the ultrasound proximity sensor will decrese the top speaker gain which
is a bad experience when this happens for a few seconds every time the
device is unlocked while playing music. The oem_pocket sensor seems to
work similar to the nonui sensor but reports 0/5 like a proximity
sensor typically does (0 is covered, 5 is not covered). It seems to use
a combination of the light sensor and orientation to report covered
when it is dark and the device is upside down (in a pocket).

Change-Id: I3df8e2513e574bcd079d18eeb3380c003780e69a
Signed-off-by: Notganesh <chimeports@gmail.com>